:fire: added a Layout and Buttons to UI

+++ src/main/java/de/hsmw/kriegZurSee/GameObjects/Boat.java
@@ -3,12 +3,13 @@ package de.hsmw.kriegZurSee.GameObjects;
 import javafx.geometry.Point2D;
 import javafx.scene.paint.Color;
 
+
 public abstract class Boat extends GameObject {
 
-    private final int length;
+    private final double length;
     private int hitPointCounter = 0;
 
-    public Boat(de.hsmw.kriegZurSee.constants.ID id, int x, int y, int width, int height, Color color) {
+    public Boat(de.hsmw.kriegZurSee.constants.ID id, int x, int y, double width, double height, Color color) {
         super(id, x, y, width, height, color);
         if (checkIsVert()) length = width / 64;
         else length = height / 64;
